fix: modify `/main` layout & add: `getStoryFeed()`
